services/market_data/requirements.txt
alpaca-py
confluent-kafka
python-dotenv
psycopg2-binary